プロジェクトの構成を変更
