--- hello-wts-batch/HelloWorldTsBatchApplication.launch
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<launchConfiguration type="org.eclipse.jdt.launching.localJavaApplication">
-<listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
-<listEntry value="/hello-wts-batch/src/main/java/jp/ne/hyukke/wts/hello/batch/HelloWorldTsBatchApplication.java"/>
-</listAttribute>
-<listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
-<listEntry value="1"/>
-</listAttribute>
-<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="jp.ne.hyukke.wts.hello.batch.HelloWorldTsBatchApplication"/>
-<stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="hello-wts-batch"/>
-<stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value="-Dspring.profiles.active=local -Dspring.batch.job.names=sampleDoSomething"/>
-</launchConfiguration>
