--- hello-wts-persistence/src/main/resources/data.sql
@@ -1,1 +1,0 @@
-insert into T_SAMPLE (SAMPLE_NAME, SAMPLE_TYPE) values ('Sample Name 1', 'TYPE_A');
