--- hello-wts-persistence/src/main/resources/persistence-local.properties
@@ -1,9 +1,0 @@
-# Database setting
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.url=jdbc:h2:file:./data/hellodb
-spring.datasource.username=sa
-spring.datasource.password=
-
-# JPA setting
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
