--- hello-wts-persistence/src/main/resources/schema.sql
@@ -1,5 +1,0 @@
-create table if not exists T_SAMPLE (
-    SAMPLE_ID int primary key auto_increment,
-    SAMPLE_NAME varchar(50),
-    SAMPLE_TYPE varchar(10)
-);
