+++ hello-wts-web/src/main/java/jp/ne/hyukke/wts/hello/core/domain/annotation/Domain.java
