+++ hello-wts-web/src/main/java/jp/ne/hyukke/wts/hello/domain/repository/SampleDomain.java
