+++ hello-wts-web/src/main/java/jp/ne/hyukke/wts/hello/persistence/entity/TSample.java
