+++ hello-wts-web/src/main/java/jp/ne/hyukke/wts/hello/persistence/repository/TSampleRepository.java
