+++ hello-wts-web/src/test/java/jp/ne/hyukke/wts/hello/web/HelloWorldTsWebApplicationTests.java
@@ -6,7 +6,7 @@ import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
-import jp.ne.hyukke.wts.hello.web.HelloWorldTsWebApplication;
+import jp.ne.hyukke.wts.hello.HelloWorldTsWebApplication;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = HelloWorldTsWebApplication.class)
