+++ template/template-domain/src/main/java/jp/ne/template/domain/DomainConfig.java
@@ -1,0 +1,13 @@
+package jp.ne.template.domain;
+
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * ドメイン層における設定クラス.
+ *
+ * @author hyukke
+ */
+@Configuration
+public class DomainConfig {
+
+}
