+++ template/template-persistence/src/main/java/jp/ne/template/persistence/PersistenceProperties.java
@@ -1,0 +1,15 @@
+package jp.ne.template.persistence;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+/**
+ * 永続層におけるプロパティクラス.
+ *
+ * @author hyukke
+ */
+@Configuration
+@PropertySource("classpath:persistence-${spring.profiles.active}.properties")
+public class PersistenceProperties {
+
+}
